examples/Keyboardio/Atreus: Use `MoveToLayer`
On the `Upper` layer, we used to have a macro that moved us back to the QWERTY
layer. It was made before we had `MoveToLayer` available. Now that we do have
it, lets use that instead of the macro.

We still keep the macro around, for the sake of compatibility, so that if anyone
who has the old macro in EEPROM, but updates the firmware, will have the key
working still.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>